requirements.txt
numpy~=1.24.3
pygame~=2.6.1
shapely~=2.0.5
networkx~=3.1
pip~=24.2
zlib~=1.2.13
wheel~=0.44.0
openssl~=3.0.15
pillow~=10.4.0
tornado~=6.4.1
setuptools~=75.1.0
matplotlib~=3.7.2
packaging~=24.1
zipp~=3.20.2
future~=0.18.3
pysocks~=1.7.1
requests~=2.32.3